requirements.txt
# HAWKSHOT Dependencies
# Install with: pip install -r requirements.txt

# DNS resolution library
dnspython>=2.4.0,<3.0.0

# HTTP requests library  
requests>=2.28.0,<3.0.0

# Cross-platform terminal colors
colorama>=0.4.6,<1.0.0

# Optional: Alternative color library (fallback)
# termcolor>=2.0.0,<3.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# flake8>=6.0.0
# black>=23.0.0
# mypy>=1.0.0
